oil_db/pip_requirements.txt
# pip requirements file for the oil_database package.
# These packages are not available via conda

pymodm>=0.4.0

pytest-raises>=0.10

git+https://github.com/NOAA-ORR-ERD/PyNUCOS.git#egg=unit_conversion
